overwrote full_fov with full_fov_phased_rx
